Don't error out when there are no clang-tidy diagnostics found

+++ scripts/run_clang_tidy.sh
@@ -2,17 +2,25 @@
 
 set -e
 
+if [ ! -d "$1" ]; then
+    echo "usage: run_clang_tidy.sh <build_dir>"
+    exit 1
+fi
+build_dir=$(readlink -f "$1")
+
 cd $(dirname $0)/..
 
 rm -Rf scripts/fixits
 mkdir -p scripts/fixits
 
-run-clang-tidy -extra-arg="-Wno-gnu-zero-variadic-macro-arguments" -j $(nproc) -config-file .clang-tidy -export-fixes scripts/fixits/fixits.yaml -use-color -p $PWD/build $PWD/src
+run-clang-tidy -extra-arg="-Wno-gnu-zero-variadic-macro-arguments" -j $(nproc) -config-file .clang-tidy -export-fixes scripts/fixits/fixits.yaml -use-color -p "$build_dir" "$PWD/src"
 
-echo "splitting fixits"
+if [ -s "scripts/fixits/fixits.yaml" ]; then
+    echo "splitting fixits"
 
-./scripts/split-clang-tidy-fixits.py scripts/fixits/fixits.yaml
+    ./scripts/split-clang-tidy-fixits.py scripts/fixits/fixits.yaml
 
-echo "fixits with auto replacements:"
+    echo "fixits with auto replacements:"
 
-grep -l Replacements:$ scripts/fixits/*/fixits.yaml | xargs dirname
+    grep -l Replacements:$ scripts/fixits/*/fixits.yaml | xargs dirname
+fi
